测试修改

+++ core/src/main/java/com/hzjq/core/worker/UnderScanShipWork.kt
@@ -20,7 +20,7 @@ class UnderScanShipWork : ScanShipCapWork {
 
     override fun onDoWorkStart(vararg args: Any) {
         caps = args[1] as MutableList<CapEntity>
-        onProgressChanged(30 + progress * (10 / 100), "正在扫描雷管信息")
+        onProgressChanged((30 + progress * (10f / 100)).toInt(), "正在扫描雷管信息")
     }
 
     override fun onScanShipSuccess(progress: CapProgressEntity) {
